mvp-services/media/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pillow==10.1.0
minio==7.2.0
